cmd: split 'surgery freelist' into separate files
