+++ cmd/bbolt/command_surgery_test.go
@@ -551,130 +551,6 @@ func testSurgeryClearPageElementsWithOverflow(t *testing.T, startIdx, endIdx int
 	compareDataAfterClearingElement(t, srcPath, output, pageId, false, startIdx, endIdx)
 }
 
-func TestSurgery_Freelist_Abandon(t *testing.T) {
-	pageSize := 4096
-	db := btesting.MustCreateDBWithOption(t, &bolt.Options{PageSize: pageSize})
-	srcPath := db.Path()
-
-	defer requireDBNoChange(t, dbData(t, srcPath), srcPath)
-
-	rootCmd := main.NewRootCommand()
-	output := filepath.Join(t.TempDir(), "db")
-	rootCmd.SetArgs([]string{
-		"surgery", "freelist", "abandon", srcPath,
-		"--output", output,
-	})
-	err := rootCmd.Execute()
-	require.NoError(t, err)
-
-	meta0 := loadMetaPage(t, output, 0)
-	assert.Equal(t, common.PgidNoFreelist, meta0.Freelist())
-	meta1 := loadMetaPage(t, output, 1)
-	assert.Equal(t, common.PgidNoFreelist, meta1.Freelist())
-}
-
-func loadMetaPage(t *testing.T, dbPath string, pageID uint64) *common.Meta {
-	_, buf, err := guts_cli.ReadPage(dbPath, 0)
-	require.NoError(t, err)
-	return common.LoadPageMeta(buf)
-}
-
-func TestSurgery_Freelist_Rebuild(t *testing.T) {
-	testCases := []struct {
-		name          string
-		hasFreelist   bool
-		expectedError error
-	}{
-		{
-			name:          "normal operation",
-			hasFreelist:   false,
-			expectedError: nil,
-		},
-		{
-			name:          "already has freelist",
-			hasFreelist:   true,
-			expectedError: main.ErrSurgeryFreelistAlreadyExist,
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			pageSize := 4096
-			db := btesting.MustCreateDBWithOption(t, &bolt.Options{
-				PageSize:       pageSize,
-				NoFreelistSync: !tc.hasFreelist,
-			})
-			srcPath := db.Path()
-
-			err := db.Update(func(tx *bolt.Tx) error {
-				// do nothing
-				return nil
-			})
-			require.NoError(t, err)
-
-			defer requireDBNoChange(t, dbData(t, srcPath), srcPath)
-
-			// Verify the freelist isn't synced in the beginning
-			meta := readMetaPage(t, srcPath)
-			if tc.hasFreelist {
-				if meta.Freelist() <= 1 || meta.Freelist() >= meta.Pgid() {
-					t.Fatalf("freelist (%d) isn't in the valid range (1, %d)", meta.Freelist(), meta.Pgid())
-				}
-			} else {
-				require.Equal(t, common.PgidNoFreelist, meta.Freelist())
-			}
-
-			// Execute `surgery freelist rebuild` command
-			rootCmd := main.NewRootCommand()
-			output := filepath.Join(t.TempDir(), "db")
-			rootCmd.SetArgs([]string{
-				"surgery", "freelist", "rebuild", srcPath,
-				"--output", output,
-			})
-			err = rootCmd.Execute()
-			require.Equal(t, tc.expectedError, err)
-
-			if tc.expectedError == nil {
-				// Verify the freelist has already been rebuilt.
-				meta = readMetaPage(t, output)
-				if meta.Freelist() <= 1 || meta.Freelist() >= meta.Pgid() {
-					t.Fatalf("freelist (%d) isn't in the valid range (1, %d)", meta.Freelist(), meta.Pgid())
-				}
-			}
-		})
-	}
-}
-
-func readMetaPage(t *testing.T, path string) *common.Meta {
-	_, activeMetaPageId, err := guts_cli.GetRootPage(path)
-	require.NoError(t, err)
-	_, buf, err := guts_cli.ReadPage(path, uint64(activeMetaPageId))
-	require.NoError(t, err)
-	return common.LoadPageMeta(buf)
-}
-
-func readPage(t *testing.T, path string, pageId int, pageSize int) []byte {
-	dbFile, err := os.Open(path)
-	require.NoError(t, err)
-	defer dbFile.Close()
-
-	fi, err := dbFile.Stat()
-	require.NoError(t, err)
-	require.GreaterOrEqual(t, fi.Size(), int64((pageId+1)*pageSize))
-
-	buf := make([]byte, pageSize)
-	byteRead, err := dbFile.ReadAt(buf, int64(pageId*pageSize))
-	require.NoError(t, err)
-	require.Equal(t, pageSize, byteRead)
-
-	return buf
-}
-
-func pageDataWithoutPageId(buf []byte) []byte {
-	return buf[8:]
-}
-
 func TestSurgeryRequiredFlags(t *testing.T) {
 	errMsgFmt := `required flag(s) "%s" not set`
 	testCases := []struct {
